+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### 📁 Other pull requests
 
 - GH-141 Fixing JournalLinkDecoder allowing to propagate 404  [\#145](https://github.com/adobe/aio-lib-java/pull/145) ([francoisledroff](https://github.com/francoisledroff))
+- GH-125 add new webhook module for payload authentication [\#144](https://github.com/adobe/aio-lib-java/pull/144) ([abhupadh](https://github.com/abhupadh))
 
 ## [aio-lib-java-1.0.6](https://github.com/adobe/aio-lib-java/tree/aio-lib-java-1.0.6) (2023-01-24)